+++ libs/execution/include/hpx/execution/execution.hpp
@@ -11,6 +11,5 @@
 #include <hpx/execution/executors/execution_parameters.hpp>
 #include <hpx/execution/executors/thread_execution.hpp>
 #include <hpx/execution/executors/thread_execution_information.hpp>
-#include <hpx/execution/executors/timed_execution.hpp>
 
 #endif
