+++ libs/executors/include/hpx/executors.hpp
@@ -17,6 +17,5 @@
 #include <hpx/executors/service_executors.hpp>
 #include <hpx/executors/thread_pool_attached_executors.hpp>
 #include <hpx/execution/executors/thread_pool_executor.hpp>
-#include <hpx/executors/timed_executors.hpp>
 
 #endif
