+++ libs/include/include/hpx/include/parallel_executors.hpp
@@ -8,8 +8,7 @@
 #define HPX_PARALLEL_EXECUTORS_MAY_11_2015_0622PM
 
 #include <hpx/executors.hpp>
-// Added here to avoid a circular dependency
-#include <hpx/parallel/executors/thread_timed_execution.hpp>
-#include <hpx/thread_executors/executors.hpp>
+#include <hpx/timed_execution.hpp>
+#include <hpx/thread_executors.hpp>
 
 #endif
