+++ libs/include/include/hpx/include/thread_executors.hpp
@@ -10,9 +10,6 @@
 #include <hpx/config.hpp>
 
 #if defined(HPX_HAVE_THREAD_EXECUTORS_COMPATIBILITY)
-// Added here to avoid a circular dependency
-#include <hpx/parallel/executors/thread_timed_execution.hpp>
-
 #include <hpx/thread_executors/current_executor.hpp>
 #include <hpx/thread_executors/default_executor.hpp>
 #include <hpx/thread_executors/embedded_thread_pool_executors.hpp>
@@ -20,6 +17,7 @@
 #include <hpx/thread_executors/service_executors.hpp>
 #include <hpx/thread_executors/thread_executor.hpp>
 #include <hpx/thread_executors/thread_pool_os_executors.hpp>
+#include <hpx/thread_executors/thread_timed_execution.hpp>
 #endif
 
 #endif
