+++ libs/timed_execution/README.rst
@@ -1,0 +1,16 @@
+
+..
+    Copyright (c) 2020 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+===============
+timed_execution
+===============
+
+This library is part of HPX.
+
+Documentation can be found `here
+<https://stellar-group.github.io/hpx-docs/latest/html/libs/timed_execution/docs/index.html>`__.
