+++ libs/timed_execution/include_compatibility/hpx/traits/is_timed_executor.hpp
@@ -5,15 +5,15 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/config.hpp>
-#include <hpx/execution/config/defines.hpp>
-#include <hpx/execution/traits/is_timed_executor.hpp>
+#include <hpx/timed_execution/config/defines.hpp>
+#include <hpx/timed_execution/traits/is_timed_executor.hpp>
 
-#if defined(HPX_EXECUTION_HAVE_DEPRECATION_WARNINGS)
+#if defined(HPX_TIMED_EXECUTION_HAVE_DEPRECATION_WARNINGS)
 #if defined(HPX_MSVC)
 #pragma message("The header hpx/traits/is_timed_executor.hpp is deprecated, \
-    please include hpx/execution/traits/is_timed_executor.hpp instead")
+    please include hpx/timed_execution/traits/is_timed_executor.hpp instead")
 #else
 #warning "The header hpx/traits/is_timed_executor.hpp is deprecated, \
-    please include hpx/execution/traits/is_timed_executor.hpp instead"
+    please include hpx/timed_execution/traits/is_timed_executor.hpp instead"
 #endif
 #endif
