+++ libs/timed_execution/tests/unit/CMakeLists.txt
@@ -1,31 +1,24 @@
-# Copyright (c) 2020 The STE||AR-Group
+# Copyright (c) 2014-2020 Hartmut Kaiser
 #
 # SPDX-License-Identifier: BSL-1.0
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 set(tests
-    sequenced_executor
-    service_executors
-  )
-
-if(HPX_WITH_THREAD_EXECUTORS_COMPATIBILITY)
-  set(tests ${tests}
-    thread_pool_attached_executors
-    )
-endif()
+    minimal_timed_async_executor
+    minimal_timed_sync_executor
+    timed_parallel_executor
+   )
 
 foreach(test ${tests})
-  set(sources
-      ${test}.cpp)
+  set(sources ${test}.cpp)
 
   set(${test}_PARAMETERS THREADS_PER_LOCALITY 4)
 
   source_group("Source Files" FILES ${sources})
 
-  set(folder_name "Tests/Unit/Modules/Executors")
+  set(folder_name "Tests/Unit/Modules/TimedExecution")
 
-  # add example executable
   add_hpx_executable(${test}_test
     INTERNAL_FLAGS
     SOURCES ${sources}
@@ -34,6 +27,5 @@ foreach(test ${tests})
     HPX_PREFIX ${HPX_BUILD_PREFIX}
     FOLDER ${folder_name})
 
-  add_hpx_unit_test("modules.executors" ${test} ${${test}_PARAMETERS})
-
+  add_hpx_unit_test("modules.timed_execution" ${test} ${${test}_PARAMETERS})
 endforeach()
