+++ libs/timed_execution/tests/unit/minimal_timed_async_executor.cpp
@@ -6,8 +6,10 @@
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
-#include <hpx/include/parallel_executors.hpp>
+#include <hpx/execution.hpp>
+#include <hpx/executors.hpp>
 #include <hpx/testing.hpp>
+#include <hpx/timed_execution.hpp>
 
 #include <algorithm>
 #include <array>
