+++ libs/timed_execution/tests/unit/minimal_timed_sync_executor.cpp
@@ -8,6 +8,7 @@
 #include <hpx/hpx_init.hpp>
 #include <hpx/include/parallel_executors.hpp>
 #include <hpx/testing.hpp>
+#include <hpx/timed_execution.hpp>
 
 #include <algorithm>
 #include <array>
